requirements.txt
asgiref==3.7.2
astroid==2.15.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
dill==0.3.7
Django==4.2.4
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
msal==1.23.0
packaging==23.1
platformdirs==3.10.0
pycparser==2.21
PyJWT==2.8.0
pylint==2.17.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.1
urllib3==2.0.4
wrapt==1.15.0
